rename opt to optimizer
