+++ query/src/sql/exec/mod.rs
@@ -31,7 +31,7 @@ use crate::sessions::QueryContext;
 use crate::sql::exec::data_schema_helper::DataSchemaHelper;
 use crate::sql::exec::expression_builder::ExpressionBuilder;
 use crate::sql::exec::util::check_physical;
-use crate::sql::opt::SExpr;
+use crate::sql::optimizer::SExpr;
 use crate::sql::Metadata;
 use crate::sql::PhysicalProject;
 use crate::sql::PhysicalScan;
